ArgosOS/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
pytesseract>=0.3.10
Pillow>=10.0.0
pdfminer.six>=20221105
python-docx>=1.1.0
openai>=1.3.7
pytest>=7.4.3
httpx>=0.25.2
python-dotenv>=1.0.0
